+++ API/Controllers/GaragesController.cs
@@ -1,11 +1,7 @@
-using System.Net.Http;
-using System.Text.Json;
-using System.Threading.Tasks;
 using API.Entities;
 using API.Data;
 using Microsoft.AspNetCore.Mvc;
-using System.Text.Json.Serialization;
-using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace API.Controllers
 {
@@ -25,27 +21,32 @@ namespace API.Controllers
         [HttpGet]
         public async Task<IActionResult> SeedData()
         {
-        using (var httpClient = new HttpClient())
-        {
-        using (var response = await httpClient.GetAsync("https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres/records"))
+            using (var httpClient = new HttpClient())
             {
-            string apiResponse = await response.Content.ReadAsStringAsync();
-            Root root = JsonConvert.DeserializeObject<Root>(apiResponse);
+                using (var response = await httpClient.GetAsync("https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres/records"))
+                {
+                    string apiResponse = await response.Content.ReadAsStringAsync();
+                    JObject jObject = JObject.Parse(apiResponse);
+                    IList<JToken> records = jObject["records"].Children().ToList();
 
-                    foreach (var record in root.Records)
+                    IEnumerable<Record> recordsList = new List<Record>();
+
+                    foreach (JToken record in records)
                     {
-                        /*if (record?.Fields != null)*/
-                        
-                            _context.Records.Add(record);
-                        
+                        //record est un objet qui contient les champs de ton JSON
+                        //Tu as link et record à l'intérieur
+
+                        //Le ToObject est un raccourci pour désérialiser le JSON en objet Record
+                        recordsList = recordsList.Append(record["record"].ToObject<Record>());
                     }
 
+                    //TODO Insérer en base les infos
+
                     await _context.SaveChangesAsync();
-
+                }
             }
-        }
 
-    return Ok();
+            return Ok();
         }
     }
 }
