+++ API/Data/DbInitializer.cs
@@ -7,33 +7,33 @@ namespace API.Data
     {
         public static void Initialize(ControleContext context)
         {
-            if(context.Field.Any()) return;
-            
-            var field = new List<Fields>
-            {
-                new Fields
-                {
-                    CctSiret = "12345678901234",
-                    CctDenomination = "Dummy Garage",
-                    CctAdresse = "123 Dummy Street",
-                    CctCodeCommune = "75001",
-                    CodePostal = "75001",
-                    CctUrl="test",
-                    CctTel = "0123456789",
-                    Region = "Île-de-France",
-                    CodeRegion = "11",
-                    Departement = "Paris",
-                    CodeDepartement = "75"
-                },
+            if(context.Records.Any()) return;
+
+            // List<Record> records = new()
+            // {
+            //     new()
+            //     {
+            //         CctSiret = "12345678901234",
+            //         CctDenomination = "Dummy Garage",
+            //         CctAdresse = "123 Dummy Street",
+            //         CctCodeCommune = "75001",
+            //         CodePostal = "75001",
+            //         CctUrl="test",
+            //         CctTel = "0123456789",
+            //         Region = "Île-de-France",
+            //         CodeRegion = "11",
+            //         Departement = "Paris",
+            //         CodeDepartement = "75"
+            //     },
                 
-            };
+            // };
             
-            //We then add the product to our Data Base context
+            // //We then add the product to our Data Base context
             
-            foreach (var fields in field)
-            {
-                context.Field.Add(fields);
-            }
+            // foreach (Record record in records)
+            // {
+            //     context.Records.Add(record);
+            // }
 
             context.SaveChanges();
         }
