+++ API/test.json
@@ -1,0 +1,1 @@
+{"total_count": 5514, "links": [{"rel": "self", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres/records?limit=10&offset=0&include_app_metas=False&include_links=False"}, {"rel": "first", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres/records?limit=10&offset=0&include_app_metas=False&include_links=False"}, {"rel": "last", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres/records?limit=10&offset=5510&include_app_metas=False&include_links=False"}, {"rel": "next", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres/records?limit=10&offset=10&include_app_metas=False&include_links=False"}], "records": [{"links": [{"rel": "self", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres/records/ced0e72d30d731902130264489b2ee966f9b2b55"}, {"rel": "datasets", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets"}, {"rel": "dataset", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres"}], "record": {"id": "ced0e72d30d731902130264489b2ee966f9b2b55", "timestamp": "2024-01-15T15:15:42.792Z", "size": 399, "fields": {"cct_siret": "30164289800037", "cat_vehicule_id": ["3", "1", "2", "5", "4"], "cat_vehicule_libelle": ["Voiture de collection", "Voiture Particuli\u00e8re", "4 x 4", "Camping-car (moins de 3,5 tonnes)", "Camionnette"], "cat_energie_id": ["2", "1"], "cat_energie_libelle": ["Diesel", "Essence"], "cct_denomination": "CENTRE DE CONTROLE TECHNIQUE FAURIE", "cct_adresse": "LE PAVE", "coordgeo": {"lon": 3.433332, "lat": 46.22081}, "cct_code_commune": "Billy", "code_postal": "03260", "cct_tel": "0470581414", "cct_url": "http://controle-technique.autosur.fr", "region": "Auvergne-Rh\u00f4ne-Alpes", "code_region": "84", "departement": "Allier", "code_departement": "03"}}}, {"links": [{"rel": "self", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres/records/1c0a088326f5e4f1e2d7f77a66b21b5d3ef4d011"}, {"rel": "datasets", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets"}, {"rel": "dataset", "href": "https://data.economie.gouv.fr/api/explore/v2.0/catalog/datasets/prix-des-controles-techniques-annuaire-des-centres"}], "record": {"id": "1c0a088326f5e4f1e2d7f77a66b21b5d3ef4d011", "timestamp": "2024-01-15T15:15:42.792Z", "size": 409, "fields": {"cct_siret": "32838567900100", "cat_vehicule_id": ["2", "5", "1", "4"], "cat_vehicule_libelle": ["4 x 4", "Camping-car (moins de 3,5 tonnes)", "Voiture Particuli\u00e8re", "Camionnette"], "cat_energie_id": ["3", "5", "1", "4", "2"], "cat_energie_libelle": ["Gaz", "Electrique", "Essence", "Hybride", "Diesel"], "cct_denomination": "CENTRE AUTO BILAN MONTREUIL", "cct_adresse": "58 BD DE L'AVENIR", "coordgeo": {"lon": 2.382019, "lat": 47.079721}, "cct_code_commune": "Bourges", "code_postal": "18000", "cct_tel": "0248204037", "cct_url": "http://controle-technique-bourges.fr", "region": "Centre-Val de Loire", "code_region": "24", "departement": "Cher", "code_departement": "18"}}}}]
